vcenter_lifecycle/requirements.txt
requests>=2.28
pyyaml>=6.0
urllib3>=1.26
